move the change_audio_config to eval_audio.py and spolish the description
